requirements.txt
python-dotenv
langchain-openai
langgraph
yfinance
langchain-core
langchain-mcp-adapters
